% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUnionsDist.R
\docType{class}
\name{allowedDist-class}
\alias{allowedDist-class}
\alias{allowedDist}
\title{An S4 union class to represent allowed dist object.}
\description{
This S4 union class describes what types of distribution are supported by the \code{optimalThreshold} package.
}
\details{
Five theoretical univariate and continuous distributions are supported by the package internally (normal, log-normal, gamma, scaled t, and logistic). However, it is possible for the user to define its own distribution type using the \code{fit} function, and passing the result to one of the two main functions: \code{trtSelThresh}, and \code{diagThresh}.
}
\seealso{
\code{\link[optimalThreshold]{fit}}, \code{\link[optimalThreshold]{trtSelThresh}}, and \code{\link[optimalThreshold]{diagThresh}}.
}
