% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDiagRelUtility.R
\docType{class}
\name{diagRelUtility-class}
\alias{diagRelUtility-class}
\alias{diagRelUtility}
\title{An S4 class to sum up the results from the decisionCurve methods.}
\description{
An S4 class to sum up the results from the decisionCurve methods.
}
\details{
You never have to create this class manually. This class is created internally when the \code{decisionCurve} method is applied to a 'diagOptThresh' object.
}
\section{Slots}{

\describe{
\item{\code{U}}{This slot is a matrix of the marker-based utility according to \code{r}.}

\item{\code{UNoTreat}}{This slot is a matrix of the utility of the 'No Treat' strategy according to \code{r}.}

\item{\code{UTreatAll}}{This slot is a matrix of the utility of the 'Treat All' strategy according to \code{r}.}

\item{\code{r}}{Risk threshold preference. Numeric argument.}
}}

\seealso{
\code{\link[optimalThreshold]{decisionCurve}} for more details on how to plot the decision curves.
}
