% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGammaDist.R
\docType{class}
\name{gammaDist-class}
\alias{gammaDist-class}
\alias{gammaDist}
\title{An S4 class to represent a gamma distribution.}
\description{
This S4 class describes the gamma distribution that is fitted to the marker values. The gamma distribution is characterized by the \code{shape} and the \code{scale} parameters.
}
\details{
You never have to create this class manually. This class is created internally when a gamma distribution is fitted on the marker values.
}
\section{Slots}{

\describe{
\item{\code{shape}}{shape parameter. Must be positive.}

\item{\code{scale}}{scale parameter. Must be strictly positive.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit a gamma distribution.
}
