% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLogisticDist.R
\docType{class}
\name{logisticDist-class}
\alias{logisticDist-class}
\alias{logisticDist}
\title{An S4 class to represent a logistic distribution.}
\description{
This S4 class describes the logistic distribution that is fitted to the marker values. The logistic distribution is characterized by the \code{location} and the \code{scale} parameters.
}
\details{
You never have to create this class manually. This class is created internally when a logistic distribution is fitted on the marker values.
}
\section{Slots}{

\describe{
\item{\code{location}}{location parameter.}

\item{\code{scale}}{scale parameter. Must be strictly positive.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit a logistic distribution.
}
