% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLogNormalDist.R
\docType{class}
\name{logNormalDist-class}
\alias{logNormalDist-class}
\alias{logNormalDist}
\title{An S4 class to represent a log-normal distribution.}
\description{
This S4 class describes the log-normal distribution that is fitted to the marker values. The log-normal distribution is characterized by the \code{mu} and the \code{sd} parameters.
}
\details{
You never have to create this class manually. This class is created internally when a log-normal distribution is fitted on the marker values.
}
\section{Slots}{

\describe{
\item{\code{mu}}{mu parameter.}

\item{\code{sd}}{standard deviation parameter. Must be strictly positive.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit a log-normal distribution.
}
