% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUnions.R
\docType{class}
\name{mcmc.listOrNull-class}
\alias{mcmc.listOrNull-class}
\alias{mcmc.listOrNull}
\title{An S4 union class to merge mcmc.list and NULL types of object.}
\description{
An S4 union class to merge mcmc.list and NULL types of object.
}
