% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassNormalDist.R
\docType{class}
\name{normalDist-class}
\alias{normalDist-class}
\alias{normalDist}
\title{An S4 class to represent a normal distribution.}
\description{
This S4 class describes the normal distribution that is fitted to the marker values. The normal distribution is characterized by the \code{mu} and the \code{sd} parameters.
}
\details{
You never have to create this class manually. This class is created internally when a normal distribution is fitted on the marker values.
}
\section{Slots}{

\describe{
\item{\code{mu}}{mu parameter.}

\item{\code{sd}}{standard deviation parameter. Must be strictly positive.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit a normal distribution.
}
