% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTrtSelOptThresh.R, R/ClassDiagOptThresh.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,trtSelOptThresh-method}
\alias{show.trtSelOptThresh}
\alias{show,diagOptThresh-method}
\alias{show.diagOptThresh}
\title{Show method}
\usage{
\S4method{show}{trtSelOptThresh}(object)

\S4method{show}{diagOptThresh}(object)
}
\arguments{
\item{object}{a \code{trtSelOptThresh} or a \code{diagOptThresh} S4 object.}
}
\value{
None
}
\description{
Show some of the slots of a \code{trtSelOptThresh} or a \code{diagOptThresh} objects.
}
