% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassStudentDist.R
\docType{class}
\name{studentDist-class}
\alias{studentDist-class}
\alias{studentDist}
\title{An S4 class to represent a scaled Student distribution.}
\description{
This S4 class describes the scaled t distribution that is fitted to the marker values. The scaled t distribution is characterized by the \code{df} (degrees of freedom), the \code{mu}, and the \code{sd} parameters.
}
\details{
You never have to create this class manually. This class is created internally when a scaled t distribution is fitted on the marker values.
}
\section{Slots}{

\describe{
\item{\code{df}}{degrees of freedom (> 0, maybe non-integer).}

\item{\code{mu}}{mu parameter.}

\item{\code{sd}}{standard deviation parameter. Must be strictly positive.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit a t distribution.
}
