% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTrtSelRelUtility.R
\docType{class}
\name{trtSelRelUtility-class}
\alias{trtSelRelUtility-class}
\alias{trtSelRelUtility}
\title{An S4 class to the results from the decisionCurve methods.}
\description{
An S4 class to the results from the decisionCurve methods.
}
\details{
You never have to create this class manually. This class is created internally when the \code{decisionCurve} method is applied to an 'optThresh' object.
}
\section{Slots}{

\describe{
\item{\code{RU}}{This slot is a matrix of the relative utility according to the r ratios.}

\item{\code{U}}{This slot is a matrix of the marker-based utility according to the r ratios.}

\item{\code{UT0}}{This slot is a matrix of the reference treatment utility according to the r ratios.}

\item{\code{UT1}}{This slot is a matrix of the innovative treatment utility according to the r ratios.}

\item{\code{Up}}{This slot is a matrix of perfect marker utility according to the r ratios.}

\item{\code{r}}{Ratio of treatment/event costs. Numeric argument.}
}}

\seealso{
\code{\link[optimalThreshold]{decisionCurve}} for more details on how to plot the decision curves.
}
