% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUndefined.R
\docType{class}
\name{undefined-class}
\alias{undefined-class}
\alias{undefined}
\title{An S4 class to represent an 'undefined' distribution.}
\description{
This class allows to fit an undefined distribution on the marker values \code{x}.
}
\details{
This class is automatically created when the user applies the \code{fit} function with the argument \code{distr="undefined"}. You never have to create manually this class, it is created internally.
}
\section{Slots}{

\describe{
\item{\code{x}}{a vector of marker values.}

\item{\code{n}}{Length of x vector (including NA values). Numeric argument.}

\item{\code{mcmc}}{This slot allows the main function to k,now whether an MCMC algorithm must be performed to sample the distribution parameters from their posterior distribution.}
}}

\seealso{
\code{\link[optimalThreshold]{fit}} for more details on how to fit an undefined distribution.
}
